Fix type annotations: replace object types with protocols, add TypeGuard for HVP narrowing
- Replace `object` parameter types with correct protocols across 10+ files
  (FunctionProtocol, GalerkinPhysicsProtocol, AbstractPhysics, etc.)
- Extend AdjointEnabledTimeSteppingResidualProtocol with missing methods
  (sensitivity_off_diag_jacobian, is_explicit, has_prev_state_hessian, native_residual)
- Extend HVPEnabledTimeSteppingResidualProtocol with prev_state_* HVP methods
- Add TypeGuard is_hvp_enabled() for safe HVP method narrowing
- Refactor TimeAdjointOperatorWithHVP to use TypeGuard instead of _has_hvp flag
- Fix no-any-return errors with proper int/float/bool wrappers
- Fix .reshape() calls to use bkd.reshape() for backend compatibility

+++ pyapprox/benchmarks/functions/multifidelity/polynomial_ensemble.py
@@ -268,7 +268,7 @@ class PolynomialEnsemble(Generic[Array]):
         cov = self.covariance(i, j)
         var_i = self._models[i].variance()
         var_j = self._models[j].variance()
-        return cov / (var_i * var_j) ** 0.5
+        return float(cov / (var_i * var_j) ** 0.5)
 
     def correlation_matrix(self) -> Array:
         """Return correlation matrix between all models.
